refactor(render): extract rendering and output formatting into dedicated crate
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
