+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "crates/agent",
     "crates/cli",
+    "crates/render",
     "crates/runtime",
     "crates/api",
     "crates/commands",
@@ -13,6 +14,7 @@ resolver = "2"
 
 [workspace.dependencies]
 agent = { path = "crates/agent" }
+render = { path = "crates/render" }
 
 [workspace.package]
 version = "0.6.4"
