+++ crates/cli/Cargo.toml
@@ -18,6 +18,7 @@ crawler = { path = "../crawler" }
 crossterm = "0.28"
 dialoguer = { version = "0.12", features = ["fuzzy-select"] }
 indicatif = "0.17"
+render = { path = "../render" }
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls"] }
 runtime = { path = "../runtime" }
 serde = { version = "1", features = ["derive"] }
@@ -26,7 +27,6 @@ sha2 = "0.10"
 tokio = { version = "1", features = ["rt-multi-thread", "time", "process"] }
 ratatui = "0.30.0"
 textwrap = "0.16"
-pulldown-cmark = "0.13"
 unicode-width = "0.2"
 zeroize = "1"
 arboard = "3"
